refactor(web): move evaluator upgrade helpers out of evals router
Relocate prepareConfigsForTemplateUpgrade and template-variable
resolution into evaluatorUpgrade.ts next to the mapping-migration
logic, merging the two near-identical variable helpers into one.
