+++ web/src/features/evals/server/evaluatorUpgrade.ts
@@ -1,9 +1,21 @@
 import {
   EvalTargetObject,
+  EvalTemplateType,
   InternalServerError,
+  LangfuseConflictError,
   observationVariableMappingList,
   variableMappingList,
 } from "@langfuse/shared";
+import { CODE_EVAL_TEMPLATE_VARIABLES } from "@/src/features/evals/utils/code-eval-template-utils";
+
+// accepts both stored templates and createTemplate inputs (CODE inputs carry no vars)
+export const getEvalTemplateVariables = (template: {
+  type: EvalTemplateType;
+  vars?: string[];
+}) =>
+  template.type === EvalTemplateType.CODE
+    ? [...CODE_EVAL_TEMPLATE_VARIABLES]
+    : (template.vars ?? []);
 
 export function prepareVariableMappingForEvaluatorUpgrade(params: {
   targetObject: string;
@@ -36,3 +48,31 @@ export function prepareVariableMappingForEvaluatorUpgrade(params: {
     missingVariables,
   };
 }
+
+export const prepareConfigsForTemplateUpgrade = (params: {
+  configs: {
+    id: string;
+    scoreName: string;
+    targetObject: string;
+    variableMapping: unknown;
+  }[];
+  nextVariables: string[];
+}) =>
+  params.configs.map((config) => {
+    const preparedMapping = prepareVariableMappingForEvaluatorUpgrade({
+      targetObject: config.targetObject,
+      variableMapping: config.variableMapping,
+      nextVariables: params.nextVariables,
+    });
+
+    if (preparedMapping.missingVariables.length > 0) {
+      throw new LangfuseConflictError(
+        `Creating a new evaluator version would invalidate the evaluator "${config.scoreName}" because it is missing mappings for new evaluator variables: ${preparedMapping.missingVariables.join(", ")}. Remove the new variable(s) from the template, or delete the evaluator "${config.scoreName}" and recreate it with a complete mapping.`,
+      );
+    }
+
+    return {
+      id: config.id,
+      variableMapping: preparedMapping.variableMapping,
+    };
+  });
